Completing first half of first quest
Finished the first half of the first quests, mind that an item has been added to the room g8, that item is coffee and is temporary and should be removed upon completion of the second half of the quest, the item was added due to test purposes. Also added a new room, storage, where you can find the book needed to talk to the npc Andars.